requirements.txt
Shapely==1.7.0
PyTurboJPEG==1.4.1
matplotlib==3.1.2
opencv_python==4.2.0.34
numpy==1.17.4
